+++ src/renderers/shaders/ShaderLib/distanceRGBA.glsl.js
@@ -4,6 +4,7 @@ export const vertex = /* glsl */`
 varying vec3 vWorldPosition;
 
 #include <common>
+#include <batching_pars_vertex>
 #include <uv_pars_vertex>
 #include <displacementmap_pars_vertex>
 #include <morphtarget_pars_vertex>
